Add support for parralel compilation.
The new default make rule does not compile libraries anymore, but directly generates the move3d executable for object files.
To compile in parralel : start make with 'make -j' or 'make -j<n>' where n is the number of desired number of compilation processes.
Hopefully, this should also solve a bug where it was necessary to do a make clean in order to compile correctly.

The former compilation rule is still available with the command 'make move3d_with_libs'.