The default RNG implementation is changed.
Move3D uses rand() from the standard C library to generate random numbers for sampling.
This causes the following problem : if a 3rd party library also makes calls to rand(),
it is impossible to use a seed to obtain reproducible results, because the order and number of calls
to rand() will vary.
This happens in particular with qt4.
The rng now used is a mersenne-twister implementation, described in MTRand.cpp/MTRand.hpp
Reference :
Mersenne Twister: A 623-Dimensionally Equidistributed
     Uniform Pseudo-Random Number Generator,
     ACM Transactions on Modeling and Computer Simulation,
     Vol. 8, No. 1, January 1998, pp. 3-30
     URL: http://www.math.sci.hiroshima-u.ac.jp/~m-mat/MT/emt.html