parametrized the distance to which moving humans cause costs
The idea is that beyond a distance, the human won't be there anyway when the robot gets there, so why add costs?
A function that reduces cost might be better, or costs that do not depend on the euclid distance, but on the map
distance to human. But euclid should be sufficient for now

set default in commit