Add a multi-thread display mode.
Multi-thread display mode is needed when using a planner thread at the same time as a drawing thread.

In this mode, the display only refers to robots using their pointers (does not use the current robot pointer).
Indeed this pointer is needed by the planner thread to set kinematic contraints (which can not be called on a robot, only by means of the current robot pointer). In addition, drawing functions that require to set the robot in a particular configuration are disabled in multi-thread mode.

Add an additional function pointer for this particular drawing functions linked to this mode.
Add a setter function for this mode.

Note, that it is temporary, the new drawing functions (robot, object, poly) have to be merged with the old in the future.